+++ src/main/java/com/example/cherrydan/oauth/repository/RefreshTokenRepository.java
@@ -1,8 +1,9 @@
 package com.example.cherrydan.oauth.repository;
 
-import com.example.cherrydan.oauth.model.RefreshToken;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
+import com.example.cherrydan.oauth.domain.RefreshToken;
 
 import java.util.Optional;
