+++ src/main/java/com/example/cherrydan/oauth/service/RefreshTokenService.java
@@ -3,7 +3,7 @@ package com.example.cherrydan.oauth.service;
 import com.example.cherrydan.common.exception.ErrorMessage;
 import com.example.cherrydan.common.exception.RefreshTokenException;
 import com.example.cherrydan.common.exception.UserException;
-import com.example.cherrydan.oauth.model.RefreshToken;
+import com.example.cherrydan.oauth.domain.RefreshToken;
 import com.example.cherrydan.oauth.repository.RefreshTokenRepository;
 import com.example.cherrydan.oauth.security.jwt.JwtTokenProvider;
 import com.example.cherrydan.user.domain.User;
